serialization tests read data from files
