+++ testdata/Model/Json/General/ReferenceArray.json
@@ -1,0 +1,1 @@
+{"melding":{"name":"testName","tags":"testTags","simple_list":{"simple_keyvalues":[{"key":"test","doubleValue":2.1,"intValue":2},{"key":"test2","doubleValue":3.1,"intValue":3}]},"toggle":true}}
