+++ testdata/Seres/Kursdomene_HvemErHvem_M_2021-04-08_5742_34627_SERES.json
@@ -1,0 +1,1 @@
+{"dataFormatProvider":"SERES","dataFormatId":"5742","dataFormatVersion":"34627","Innrapportoer":{"geek":{"navn":"Ronny","foedselsdato":"1971-11-02","epost":"ronny.birkeli@gmail.com"}},"InnrapporterteData":{"geekType":"backend","altinnErfaringAAr":0}}
